requirements.txt
fastapi
uvicorn
Pillow
numpy
cairosvg
